Move user related entities to separate package. Cleanup event entities a bit.
