+++ ippon/permissions.py
@@ -1,10 +1,4 @@
-from django.views.generic.base import View
-from rest_framework import permissions
-from rest_framework.request import Request
-
 import ippon.models.tournament as tm
-
-from ippon.models.event import Event, EventAdmin
 
 
 def is_user_admin_of_the_tournament(request, tournament):
@@ -34,18 +28,3 @@ def get_tournament_from_fight(fight):
     return fight.team_fight.tournament
 
 
-class IsEventOwnerOrReadOnly(permissions.BasePermission):
-    def has_permission(self, request: Request, view: View) -> bool:
-        if request.method in permissions.SAFE_METHODS:
-            return True
-        elif request.method == "POST":
-            try:
-                return request.user.pk == int(request.data["event_owner"])
-            except (KeyError, ValueError):
-                return False
-        else:
-            pk = view.kwargs["pk"]
-            try:
-                return any([i.user == request.user for i in EventAdmin.objects.filter(event=Event.objects.get(pk=pk))])
-            except Exception:
-                return False
