+++ ippon/user/serailzers.py
@@ -1,0 +1,26 @@
+from django.contrib.auth.models import User
+from rest_framework import serializers
+
+
+class UserRegistrationSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = (
+            'username',
+            'email',
+            'password'
+        )
+        extra_kwargs = {
+            'password': {
+                'write_only': True,
+            },
+        }
+
+
+class MinimalUserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = (
+            'id',
+            'username'
+        )
